+++ src/main/kotlin/net/fallingangel/jimmerdto/psi/DTOTokenTypes.kt
@@ -1,5 +1,7 @@
 package net.fallingangel.jimmerdto.psi
 
+import com.intellij.psi.tree.TokenSet
+
 object DTOTokenTypes {
     @JvmField
     val LINE_COMMENT = DTOTokenType("LINE_COMMENT")
@@ -9,4 +11,6 @@ object DTOTokenTypes {
 
     @JvmField
     val DOC_COMMENT = DTOTokenType("DOC_COMMENT")
+
+    val comments = TokenSet.create(LINE_COMMENT, BLOCK_COMMENT, DOC_COMMENT)
 }