+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Implement `PsiReferenceContributor`, complete the reference association between dto prop and entity prop
 * Reference parsing of imported elements is supported
+* Blank lines after comments are formatted
 
 ### Fixed